a los usuarios específicos por parametros, se les añade acceso a la base de datos en get, patch, delete, con los metodos de Astro db

+++ src/pages/api/clients/[clientId].ts
@@ -1,5 +1,5 @@
 import type { APIRoute, GetStaticPaths } from "astro";
-import { Clients, db } from "astro:db";
+import { Clients, db, eq } from "astro:db";
 
 export const prerender = false;
 
@@ -7,21 +7,16 @@ export const prerender = false;
 // en este caso se están generando las rutas con los params solicitados al servido
 export const GET: APIRoute = async ({ params, request }) => {
 	// entre corchetes puedo usar cualquier nombre, el que mejor describa el parámetro que esperas en la URL
-	const { clientId } = params;
+	const clientId = params.clientId ?? "";
 
-	return new Response(
-		JSON.stringify({
-			method: "GET",
-			clientId: 1,
-			slug: clientId,
-		}),
-		{
-			status: 200,
-			headers: {
-				"Content-Type": "application/json",
-			},
-		}
-	);
+	const user = await db.select().from(Clients).where(eq(Clients.id, +clientId));
+
+	return new Response(JSON.stringify(user), {
+		status: 200,
+		headers: {
+			"Content-Type": "application/json",
+		},
+	});
 };
 
 // POST
@@ -68,28 +63,27 @@ export const PUT: APIRoute = async ({ params, request }) => {
 
 // PATCH
 export const PATCH: APIRoute = async ({ params, request }) => {
-    const { clientId } = params;
+	const clientId = params.clientId ?? "";
 
 	try {
 		const { id, ...body } = await request.json();
 
-		const resp = await db.insert(Clients).values(body);
-		console.log("🚀 ~ constPOST:APIRoute= ~ resp:", resp);
-		const { lastInsertRowid } = resp;
+		const results = await db
+			.update(Clients)
+			.set(body)
+			.where(eq(Clients.id, +clientId));
 
-		return new Response(
-			JSON.stringify({
-				method: "POST",
-				id: +lastInsertRowid!.toString(),
-				...body,
-			}),
-			{
-				status: 200,
-				headers: {
-					"Content-Type": "application/json",
-				},
-			}
-		);
+		const updatedClient = await db
+			.select()
+			.from(Clients)
+			.where(eq(Clients.id, +clientId));
+
+		return new Response(JSON.stringify(updatedClient.at(0)), {
+			status: 200,
+			headers: {
+				"Content-Type": "application/json",
+			},
+		});
 	} catch (error) {
 		console.error("🚀 ~ constPOST:APIRoute= ~ error:", error);
 		return new Response(JSON.stringify({ msg: "No body found" }), {
@@ -118,19 +112,39 @@ export const PATCH: APIRoute = async ({ params, request }) => {
 
 // DELETE
 export const DELETE: APIRoute = async ({ params, request }) => {
-	const { clientId } = params;
+	const clientId = params.clientId ?? "";
 
-	return new Response(
-		JSON.stringify({
-			method: "DELETE",
-			clientId: 1,
-			slug: clientId,
-		}),
-		{
-			status: 200,
-			headers: {
-				"Content-Type": "application/json",
-			},
-		}
-	);
+	const { rowsAffected } = await db
+		.delete(Clients)
+		.where(eq(Clients.id, +clientId));
+
+	if (rowsAffected > 0) {
+		return new Response(
+			JSON.stringify({
+				method: "DELETE",
+				clientId: clientId,
+				msg: "Client deleted",
+			}),
+			{
+				status: 200,
+				headers: {
+					"Content-Type": "application/json",
+				},
+			}
+		);
+	}
+
+    return new Response(
+        JSON.stringify({
+            method: "DELETE",
+            clientId: clientId,
+            msg: `Client ${clientId} not found`,
+        }),
+        {
+            status: 404,
+            headers: {
+                "Content-Type": "application/json",
+            },
+        }
+    );
 };
